pyproject.toml
[project]
name = "renamerapp"
version = "0.1.0"
description = "Google Drive Image Renamer (Increment 1)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "keyring>=25.0.0",
    "pytest>=8.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.7.0",
    "streamlit[pdf]>=1.30.0",
    "python-dotenv>=1.0.1",
]
